+++ PayPal/src/main/java/com/braintreepayments/api/PayPalPaymentAuthCallback.java
@@ -1,0 +1,17 @@
+package com.braintreepayments.api;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentActivity;
+
+/**
+ * Callback for receiving result of
+ * {@link PayPalClient#createPaymentAuthRequest(FragmentActivity, PayPalRequest, PayPalPaymentAuthCallback)}.
+ */
+public interface PayPalPaymentAuthCallback {
+
+    /**
+     * @param paymentAuthRequest a request used to launch the PayPal web authentication flow
+     * @param error          an exception that occurred while initiating a PayPal transaction
+     */
+    void onResult(PayPalPaymentAuthRequest paymentAuthRequest, @Nullable Exception error);
+}