+++ PayPal/src/test/java/com/braintreepayments/api/PayPalClientUnitTest.java
@@ -33,8 +33,8 @@ public class PayPalClientUnitTest {
     private Configuration payPalEnabledConfig;
     private Configuration payPalDisabledConfig;
 
-    private PayPalBrowserSwitchResultCallback payPalBrowserSwitchResultCallback;
-    private PayPalFlowStartedCallback payPalFlowStartedCallback;
+    private PayPalTokenizeCallback payPalTokenizeCallback;
+    private PayPalPaymentAuthCallback paymentAuthCallback;
 
     @Before
     public void beforeEach() throws JSONException {
@@ -43,21 +43,21 @@ public class PayPalClientUnitTest {
         payPalEnabledConfig = Configuration.fromJson(Fixtures.CONFIGURATION_WITH_LIVE_PAYPAL);
         payPalDisabledConfig = Configuration.fromJson(Fixtures.CONFIGURATION_WITH_DISABLED_PAYPAL);
 
-        payPalBrowserSwitchResultCallback = mock(PayPalBrowserSwitchResultCallback.class);
-        payPalFlowStartedCallback = mock(PayPalFlowStartedCallback.class);
+        payPalTokenizeCallback = mock(PayPalTokenizeCallback.class);
+        paymentAuthCallback = mock(PayPalPaymentAuthCallback.class);
     }
 
     @Test
-    public void tokenizePayPalAccount_callsBackPayPalResponse() throws JSONException {
+    public void createPaymentAuthRequest_callsBackPayPalResponse() throws JSONException {
         PayPalVaultRequest payPalVaultRequest = new PayPalVaultRequest();
         payPalVaultRequest.setMerchantAccountId("sample-merchant-account-id");
 
-        PayPalResponse payPalResponse = new PayPalResponse(payPalVaultRequest)
+        PayPalPaymentAuthRequest paymentAuthRequest = new PayPalPaymentAuthRequest(payPalVaultRequest)
                 .approvalUrl("https://example.com/approval/url")
                 .successUrl("https://example.com/success/url")
                 .clientMetadataId("sample-client-metadata-id");
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder()
-                .sendRequestSuccess(payPalResponse)
+                .sendRequestSuccess(paymentAuthRequest)
                 .build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -66,15 +66,15 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, payPalVaultRequest, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, payPalVaultRequest, paymentAuthCallback);
 
-        ArgumentCaptor<PayPalResponse> captor =
-                ArgumentCaptor.forClass(PayPalResponse.class);
-        verify(payPalFlowStartedCallback).onResult(captor.capture(), isNull());
+        ArgumentCaptor<PayPalPaymentAuthRequest> captor =
+                ArgumentCaptor.forClass(PayPalPaymentAuthRequest.class);
+        verify(paymentAuthCallback).onResult(captor.capture(), isNull());
 
-        PayPalResponse payPalResponseResult = captor.getValue();
+        PayPalPaymentAuthRequest paymentAuthRequestCaptured = captor.getValue();
 
-        BrowserSwitchOptions browserSwitchOptions = payPalResponseResult.getBrowserSwitchOptions();
+        BrowserSwitchOptions browserSwitchOptions = paymentAuthRequestCaptured.getBrowserSwitchOptions();
         assertEquals(BraintreeRequestCodes.PAYPAL, browserSwitchOptions.getRequestCode());
         assertFalse(browserSwitchOptions.isLaunchAsNewTask());
 
@@ -90,16 +90,16 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void tokenizePayPalAccount_whenLaunchesBrowserSwitchAsNewTaskEnabled_setsNewTaskOption() {
+    public void createPaymentAuthRequest_whenLaunchesBrowserSwitchAsNewTaskEnabled_setsNewTaskOption() {
         PayPalVaultRequest payPalVaultRequest = new PayPalVaultRequest();
         payPalVaultRequest.setMerchantAccountId("sample-merchant-account-id");
 
-        PayPalResponse payPalResponse = new PayPalResponse(payPalVaultRequest)
+        PayPalPaymentAuthRequest paymentAuthRequest = new PayPalPaymentAuthRequest(payPalVaultRequest)
                 .approvalUrl("https://example.com/approval/url")
                 .successUrl("https://example.com/success/url")
                 .clientMetadataId("sample-client-metadata-id");
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder()
-                .sendRequestSuccess(payPalResponse)
+                .sendRequestSuccess(paymentAuthRequest)
                 .build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -109,27 +109,27 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, payPalVaultRequest, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, payPalVaultRequest, paymentAuthCallback);
 
-        ArgumentCaptor<PayPalResponse> captor =
-                ArgumentCaptor.forClass(PayPalResponse.class);
-        verify(payPalFlowStartedCallback).onResult(captor.capture(), isNull());
+        ArgumentCaptor<PayPalPaymentAuthRequest> captor =
+                ArgumentCaptor.forClass(PayPalPaymentAuthRequest.class);
+        verify(paymentAuthCallback).onResult(captor.capture(), isNull());
 
-        PayPalResponse payPalResponseResult = captor.getValue();
-        assertTrue(payPalResponseResult.getBrowserSwitchOptions().isLaunchAsNewTask());
+        PayPalPaymentAuthRequest paymentAuthRequestCaptured = captor.getValue();
+        assertTrue(paymentAuthRequestCaptured.getBrowserSwitchOptions().isLaunchAsNewTask());
     }
 
     @Test
-    public void tokenizePayPalAccount_whenVaultRequest_sendsAnalyticsEvents() {
+    public void createPaymentAuthRequest_whenVaultRequest_sendsAnalyticsEvents() {
         PayPalVaultRequest payPalVaultRequest = new PayPalVaultRequest();
         payPalVaultRequest.setMerchantAccountId("sample-merchant-account-id");
 
-        PayPalResponse payPalResponse = new PayPalResponse(payPalVaultRequest)
+        PayPalPaymentAuthRequest paymentAuthRequest = new PayPalPaymentAuthRequest(payPalVaultRequest)
                 .approvalUrl("https://example.com/approval/url")
                 .successUrl("https://example.com/success/url")
                 .clientMetadataId("sample-client-metadata-id");
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder()
-                .sendRequestSuccess(payPalResponse)
+                .sendRequestSuccess(paymentAuthRequest)
                 .build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -138,7 +138,7 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, payPalVaultRequest, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, payPalVaultRequest, paymentAuthCallback);
 
         verify(braintreeClient).sendAnalyticsEvent("paypal.billing-agreement.selected");
         verify(braintreeClient).sendAnalyticsEvent(
@@ -146,7 +146,7 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void tokenizePayPalAccount_whenPayPalNotEnabled_returnsError() {
+    public void createPaymentAuthRequest_whenPayPalNotEnabled_returnsError() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -155,11 +155,11 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, new PayPalCheckoutRequest("1.00"),
-                payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, new PayPalCheckoutRequest("1.00"),
+                paymentAuthCallback);
 
         ArgumentCaptor<Exception> errorCaptor = ArgumentCaptor.forClass(Exception.class);
-        verify(payPalFlowStartedCallback).onResult(isNull(), errorCaptor.capture());
+        verify(paymentAuthCallback).onResult(isNull(), errorCaptor.capture());
         assertTrue(errorCaptor.getValue() instanceof BraintreeException);
         assertEquals("PayPal is not enabled. " +
                 "See https://developer.paypal.com/braintree/docs/guides/paypal/overview/android/v4 " +
@@ -167,7 +167,7 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void tokenizePayPalAccount_whenDeviceCantPerformBrowserSwitch_returnsError() {
+    public void createPaymentAuthRequest_whenDeviceCantPerformBrowserSwitch_returnsError() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -177,11 +177,11 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, new PayPalCheckoutRequest("1.00"),
-                payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, new PayPalCheckoutRequest("1.00"),
+                paymentAuthCallback);
 
         ArgumentCaptor<Exception> errorCaptor = ArgumentCaptor.forClass(Exception.class);
-        verify(payPalFlowStartedCallback).onResult(isNull(), errorCaptor.capture());
+        verify(paymentAuthCallback).onResult(isNull(), errorCaptor.capture());
         assertTrue(errorCaptor.getValue() instanceof BraintreeException);
         assertEquals("AndroidManifest.xml is incorrectly configured or another app " +
                         "defines the same browser switch url as this app. See " +
@@ -192,17 +192,17 @@ public class PayPalClientUnitTest {
 
 
     @Test
-    public void tokenizePayPalAccount_whenCheckoutRequest_sendsBrowserSwitchStartAnalyticsEvent() {
+    public void createPaymentAuthRequest_whenCheckoutRequest_sendsBrowserSwitchStartAnalyticsEvent() {
         PayPalCheckoutRequest payPalCheckoutRequest = new PayPalCheckoutRequest("1.00");
         payPalCheckoutRequest.setIntent("authorize");
         payPalCheckoutRequest.setMerchantAccountId("sample-merchant-account-id");
 
-        PayPalResponse payPalResponse = new PayPalResponse(payPalCheckoutRequest)
+        PayPalPaymentAuthRequest paymentAuthRequest = new PayPalPaymentAuthRequest(payPalCheckoutRequest)
                 .approvalUrl("https://example.com/approval/url")
                 .successUrl("https://example.com/success/url")
                 .clientMetadataId("sample-client-metadata-id");
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder()
-                .sendRequestSuccess(payPalResponse)
+                .sendRequestSuccess(paymentAuthRequest)
                 .build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -211,14 +211,14 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, payPalCheckoutRequest, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, payPalCheckoutRequest, paymentAuthCallback);
 
         verify(braintreeClient).sendAnalyticsEvent("paypal.single-payment.selected");
         verify(braintreeClient).sendAnalyticsEvent("paypal.single-payment.browser-switch.started");
     }
 
     @Test
-    public void tokenizePayPalAccount_whenCheckoutRequest_sendsPayPalPayLaterOfferedAnalyticsEvent() {
+    public void createPaymentAuthRequest_whenCheckoutRequest_sendsPayPalPayLaterOfferedAnalyticsEvent() {
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
@@ -226,13 +226,13 @@ public class PayPalClientUnitTest {
                 new PayPalClient(braintreeClient, payPalInternalClient);
         PayPalCheckoutRequest request = new PayPalCheckoutRequest("1.00");
         request.setShouldOfferPayLater(true);
-        sut.tokenizePayPalAccount(activity, request, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, request, paymentAuthCallback);
 
         verify(braintreeClient).sendAnalyticsEvent("paypal.single-payment.paylater.offered");
     }
 
     @Test
-    public void tokenizePayPalAccount_whenVaultRequest_sendsPayPalRequestViaInternalClient() {
+    public void createPaymentAuthRequest_whenVaultRequest_sendsPayPalRequestViaInternalClient() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -243,14 +243,14 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, payPalRequest, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, payPalRequest, paymentAuthCallback);
 
         verify(payPalInternalClient).sendRequest(same(activity), same(payPalRequest),
                 any(PayPalInternalClientCallback.class));
     }
 
     @Test
-    public void tokenizePayPalAccount_whenCheckoutRequest_sendsPayPalRequestViaInternalClient() {
+    public void createPaymentAuthRequest_whenCheckoutRequest_sendsPayPalRequestViaInternalClient() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder()
@@ -261,14 +261,14 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, payPalRequest, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, payPalRequest, paymentAuthCallback);
 
         verify(payPalInternalClient).sendRequest(same(activity), same(payPalRequest),
                 any(PayPalInternalClientCallback.class));
     }
 
     @Test
-    public void tokenizePayPalAccount_sendsPayPalCreditOfferedAnalyticsEvent() {
+    public void createPaymentAuthRequest_sendsPayPalCreditOfferedAnalyticsEvent() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
 
@@ -277,30 +277,30 @@ public class PayPalClientUnitTest {
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
-        sut.tokenizePayPalAccount(activity, payPalRequest, payPalFlowStartedCallback);
+        sut.createPaymentAuthRequest(activity, payPalRequest, paymentAuthCallback);
 
         verify(braintreeClient).sendAnalyticsEvent("paypal.billing-agreement.credit.offered");
     }
 
     @Test
-    public void onBrowserSwitchResult_whenResultNull_callsBackError() {
+    public void tokenize_whenResultNull_callsBackError() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(null, payPalBrowserSwitchResultCallback);
+        sut.tokenize(null, payPalTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(payPalBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(payPalTokenizeCallback).onResult(isNull(), captor.capture());
 
         assertNotNull(captor.getValue());
         assertEquals("PayPalBrowserSwitchResult cannot be null", captor.getValue().getMessage());
     }
 
     @Test
-    public void onBrowserSwitchResult_whenErrorNotNull_callsBackError() {
+    public void tokenize_whenErrorNotNull_callsBackError() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
 
@@ -308,31 +308,31 @@ public class PayPalClientUnitTest {
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
         Exception expectedError = new Exception("error");
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(expectedError);
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(expectedError);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(payPalBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(payPalTokenizeCallback).onResult(isNull(), captor.capture());
 
         assertNotNull(captor.getValue());
         assertSame(expectedError, captor.getValue());
     }
 
     @Test
-    public void onBrowserSwitchResult_whenResultAndErrorNull_callsBackUnexpectedError() {
+    public void tokenize_whenResultAndErrorNull_callsBackUnexpectedError() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
 
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult = new PayPalBrowserSwitchResult(
+        PayPalPaymentAuthResult payPalPaymentAuthResult = new PayPalPaymentAuthResult(
                 (BrowserSwitchResult) null);
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(payPalBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(payPalTokenizeCallback).onResult(isNull(), captor.capture());
 
         assertNotNull(captor.getValue());
         assertEquals("An unexpected error occurred", captor.getValue().getMessage());
@@ -340,7 +340,7 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void onBrowserSwitchResult_withBillingAgreement_tokenizesResponseOnSuccess()
+    public void tokenize_withBillingAgreement_tokenizesResponseOnSuccess()
             throws JSONException {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
@@ -362,17 +362,17 @@ public class PayPalClientUnitTest {
         Uri uri = Uri.parse(approvalUrl);
         when(browserSwitchResult.getDeepLinkUrl()).thenReturn(uri);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         ArgumentCaptor<PayPalAccount> captor = ArgumentCaptor.forClass(PayPalAccount.class);
         verify(payPalInternalClient).tokenize(captor.capture(),
-                any(PayPalBrowserSwitchResultCallback.class));
+                any(PayPalTokenizeCallback.class));
 
         PayPalAccount payPalAccount = captor.getValue();
         JSONObject tokenizePayload = payPalAccount.buildJSON();
@@ -391,7 +391,7 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void onBrowserSwitchResult_withOneTimePayment_tokenizesResponseOnSuccess()
+    public void tokenize_withOneTimePayment_tokenizesResponseOnSuccess()
             throws JSONException {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
@@ -413,17 +413,17 @@ public class PayPalClientUnitTest {
         Uri uri = Uri.parse(approvalUrl);
         when(browserSwitchResult.getDeepLinkUrl()).thenReturn(uri);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         ArgumentCaptor<PayPalAccount> captor = ArgumentCaptor.forClass(PayPalAccount.class);
         verify(payPalInternalClient).tokenize(captor.capture(),
-                any(PayPalBrowserSwitchResultCallback.class));
+                any(PayPalTokenizeCallback.class));
 
         PayPalAccount payPalAccount = captor.getValue();
         JSONObject tokenizePayload = payPalAccount.buildJSON();
@@ -445,7 +445,7 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void onBrowserSwitchResult_withBillingAgreement_sendsAnalyticsEvents()
+    public void tokenize_withBillingAgreement_sendsAnalyticsEvents()
             throws JSONException {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
@@ -467,20 +467,20 @@ public class PayPalClientUnitTest {
         Uri uri = Uri.parse(approvalUrl);
         when(browserSwitchResult.getDeepLinkUrl()).thenReturn(uri);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         verify(braintreeClient).sendAnalyticsEvent(
                 "paypal.billing-agreement.browser-switch.succeeded");
     }
 
     @Test
-    public void onBrowserSwitchResult_oneTimePayment_sendsAnalyticsEvents() throws JSONException {
+    public void tokenize_oneTimePayment_sendsAnalyticsEvents() throws JSONException {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
         String approvalUrl =
@@ -501,20 +501,20 @@ public class PayPalClientUnitTest {
         Uri uri = Uri.parse(approvalUrl);
         when(browserSwitchResult.getDeepLinkUrl()).thenReturn(uri);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         verify(braintreeClient).sendAnalyticsEvent(
                 "paypal.single-payment.browser-switch.succeeded");
     }
 
     @Test
-    public void onBrowserSwitchResult_whenPayPalCreditPresent_sendsAnalyticsEvents()
+    public void tokenize_whenPayPalCreditPresent_sendsAnalyticsEvents()
             throws JSONException {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder()
                 .tokenizeSuccess(PayPalAccountNonce.fromJSON(
@@ -539,19 +539,19 @@ public class PayPalClientUnitTest {
         Uri uri = Uri.parse(approvalUrl);
         when(browserSwitchResult.getDeepLinkUrl()).thenReturn(uri);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         verify(braintreeClient).sendAnalyticsEvent("paypal.credit.accepted");
     }
 
     @Test
-    public void onBrowserSwitchResult_whenCancelUriReceived_notifiesCancellationAndSendsAnalyticsEvent()
+    public void tokenize_whenCancelUriReceived_notifiesCancellationAndSendsAnalyticsEvent()
             throws JSONException {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
 
@@ -572,16 +572,16 @@ public class PayPalClientUnitTest {
         Uri uri = Uri.parse(approvalUrl);
         when(browserSwitchResult.getDeepLinkUrl()).thenReturn(uri);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(payPalBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(payPalTokenizeCallback).onResult(isNull(), captor.capture());
 
         Exception exception = captor.getValue();
         assertTrue(exception instanceof UserCanceledException);
@@ -593,23 +593,23 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void onBrowserSwitchResult_whenBrowserSwitchCanceled_forwardsExceptionAndSendsAnalyticsEvent() {
+    public void tokenize_whenBrowserSwitchCanceled_forwardsExceptionAndSendsAnalyticsEvent() {
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder().build();
         BrowserSwitchResult browserSwitchResult = mock(BrowserSwitchResult.class);
         when(browserSwitchResult.getStatus()).thenReturn(BrowserSwitchStatus.CANCELED);
 
         when(browserSwitchResult.getRequestMetadata()).thenReturn(new JSONObject());
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
 
         ArgumentCaptor<Exception> captor = ArgumentCaptor.forClass(Exception.class);
-        verify(payPalBrowserSwitchResultCallback).onResult(isNull(), captor.capture());
+        verify(payPalTokenizeCallback).onResult(isNull(), captor.capture());
 
         Exception exception = captor.getValue();
         assertTrue(exception instanceof UserCanceledException);
@@ -621,7 +621,7 @@ public class PayPalClientUnitTest {
     }
 
     @Test
-    public void onBrowserSwitchResult_whenPayPalInternalClientTokenizeResult_callsBackResult()
+    public void tokenize_whenPayPalInternalClientTokenizeResult_callsBackResult()
             throws JSONException {
         PayPalAccountNonce payPalAccountNonce = mock(PayPalAccountNonce.class);
         PayPalInternalClient payPalInternalClient = new MockPayPalInternalClientBuilder()
@@ -646,13 +646,13 @@ public class PayPalClientUnitTest {
         Uri uri = Uri.parse(approvalUrl);
         when(browserSwitchResult.getDeepLinkUrl()).thenReturn(uri);
 
-        PayPalBrowserSwitchResult payPalBrowserSwitchResult =
-                new PayPalBrowserSwitchResult(browserSwitchResult);
+        PayPalPaymentAuthResult payPalPaymentAuthResult =
+                new PayPalPaymentAuthResult(browserSwitchResult);
         BraintreeClient braintreeClient = new MockBraintreeClientBuilder().build();
         PayPalClient sut =
                 new PayPalClient(braintreeClient, payPalInternalClient);
 
-        sut.onBrowserSwitchResult(payPalBrowserSwitchResult, payPalBrowserSwitchResultCallback);
-        verify(payPalBrowserSwitchResultCallback).onResult(same(payPalAccountNonce), isNull());
+        sut.tokenize(payPalPaymentAuthResult, payPalTokenizeCallback);
+        verify(payPalTokenizeCallback).onResult(same(payPalAccountNonce), isNull());
     }
 }